WIP: More actions taken to establish clearer names / responsibilities between components
